pyproject.toml
[tool.poetry]
name = "discord-bottibot"
version = "0.1.0"
description = "ボッチ用bot"
authors = ["poro <tamagoporo0920@yahoo.co.jp>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "discord_bottibot"}]

[tool.poetry.dependencies]
python = "^3.8"
discord-py = "^2.1.1"
openai = "^0.26.5"
python-dotenv = "^0.21.1"
google-cloud-logging = "^3.5.0"
google-cloud-storage = "^2.7.0"
flask = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
